Refactor to new base class
