+++ src/index.ts
@@ -2,7 +2,7 @@ import SegfaultHandler from 'segfault-handler';
 
 export * from './launchpads/MK2/LaunchpadMK2.js';
 export * from './launchpads/MK3/LaunchpadMK3.js';
-export * from './launchpads/BaseLaunchpad.js';
+export * from './launchpads/base/BaseLaunchpad.js';
 export * as colors from './colorHelpers.js';
 export * as utils from './utils.js';
